Improve 64-bit alignment cleanliness
git-svn-id: https://llvm.org/svn/llvm-project/poolalloc/trunk@17417 91177308-0d34-0410-b5e6-96231b3b80d8

+++ runtime/FL2Allocator/PoolAllocator.h
@@ -19,10 +19,9 @@ struct PoolSlab;
 struct FreedNodeHeader;
 
 // NodeHeader - Each block of memory is preceeded in the the pool by one of
-// these headers.  If the node is allocated, the ObjectSize value is used, if
-// the object is free, the 'Next' value is used.
+// these headers.
 struct NodeHeader {
-  unsigned Size;
+  unsigned long Size;
 };
